pyproject.toml
[project]
name = "talos-landing"
version = "0.1.0"
description = "Landing page for taloslab.cc — live Kubernetes metrics & infrastructure showcase"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1",
    "gunicorn>=23.0",
    "kubernetes>=31.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
